Fix more "implicit conversion shortens 64-bit value into a 32-bit value" warnings via casts changing variable types.  Also remove a couple of pointless statements.
* subversion/libsvn_subr/subst.c
  (keyword_printf),
* subversion/libsvn_subr/utf.c
  (invalid_utf8),
* subversion/libsvn_subr/opt.c
  (svn_opt_parse_path, svn_opt__split_arg_at_peg_revision),
* subversion/libsvn_subr/dirent_uri.c
  (svn_uri_join),
* subversion/libsvn_subr/io.c
  (svn_io_read_link, svn_io_detect_mimetype2),
* subversion/libsvn_subr/mergeinfo.c
  (svn_mergeinfo__remove_prefix_from_catalog),
* subversion/libsvn_subr/path.c
  (svn_path_is_single_path_component),
* subversion/libsvn_delta/svndiff.c
  (zlib_encode):
    Update variable types.

* subversion/libsvn_subr/sqlite.c
  (svn_sqlite__bind_blob),
* subversion/libsvn_subr/macos_keychain.c
  (keychain_password_set, keychain_password_get)
    Add an explicit cast.

* subversion/libsvn_subr/io.c
  (cstring_to_utf8, cstring_from_utf8):  Remove unused statements.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@879100 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_subr/mergeinfo.c
@@ -1661,7 +1661,7 @@ svn_mergeinfo__remove_prefix_from_catalog(svn_mergeinfo_catalog_t *out_catalog,
                                           apr_pool_t *pool)
 {
   apr_hash_index_t *hi;
-  int prefix_len = strlen(prefix);
+  size_t prefix_len = strlen(prefix);
 
   *out_catalog = apr_hash_make(pool);
